Created a class ArrayListAndLinkedList, which contains two methods:
processArrayList: This method takes an ArrayList of strings as input and performs the following actions:
Removes all elements containing the letter 'a'.
Adds the string "NewElement" to the end of the list.
Displays all elements of the list.
processLinkedList: This method takes a LinkedList of numbers as input and performs the following actions:
Doubles the value of each element.
Removes the first three elements of the list.
Displays all elements of the list.
In the main method, create an ArrayList and LinkedList objects, add a few elements to each collection, and call the appropriate processArrayList and processLinkedList methods to process them.