The Student class was created following the rules:
name (student name) of type String.
grade (student grade) of type int.
Created a TreeMap named StudentMap, where the key is the student's name and the value is the student's grade.

Wrote the addStudent method to add a new student to the TreeMap. The method takes the student's name and his measuring instrument and writes the corresponding entry into the TreeMap.

Wrote the getTopStudent method to get the name of the first student from TreeMap.

Wrote a printStudentMap method to calculate the TreeMap score in the console.

Created a protest class object and added some students, made a list of students and got the first student.