Implemented the Schoolboy class to represent information about schoolchildren, as well as the SchoolboyList class, which represents a list of schoolchildren and has the ability to iterate over it using an iterator.
Implemented the Schoolboy class with the name and age fields, as well as the SchoolboyList class, which is a list of Schoolboy objects.

The SchoolboyList class implements the Iterable<Schoolboy> interface so that its objects can be used in a for-each loop.
Also in this class there is a method addSchoolboy(Schoolboy schoolboy), which adds a student to the list.

In addition, in the SchoolboyList class I implemented the nested SchoolboyIterator class, which implements the Iterator<Schoolboy> interface.
This iterator allows you to iterate through a list of students, starting from the first element to the last.