The producer thread puts 1000 elements into the list. If the queue is full, then I call wait() until it becomes free. After the element is added, I call notify() on it so that the consuming thread waiting for the elements to be freed knows that there is already something to consume.
The consumer thread runs in daemon mode under while(true) and stops if it receives a poison pill (it's just the last element).
As in the first case, if the queue is empty, then wait() is called until the situation changes.
After the element is used, I call notify() on it so that the producer thread (perhaps waiting for the queue to fill) knows that there is now free space to add at least one element.